requirements.txt
flask==3.0.2
flask-cors==4.0.1
werkzeug==3.0.3
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.5.0
joblib==1.4.2
python-dotenv==1.0.1
supabase==2.10.0
postgrest==0.18.0
tqdm>=4.66.0
gunicorn==23.0.0
